Extract remove site action modal and rename use-site-actions back to TS
