+++ client/a8c-for-agencies/sections/sites/site-actions/use-site-actions.ts
@@ -1,24 +1,22 @@
 import page from '@automattic/calypso-router';
-import { Button, __experimentalHStack as HStack } from '@wordpress/components';
 import { external, trash } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
 import { useContext, useMemo } from 'react';
 import { DATAVIEWS_LIST } from 'calypso/a8c-for-agencies/components/items-dashboard/constants';
 import { DataViewsState } from 'calypso/a8c-for-agencies/components/items-dashboard/items-dataviews/interfaces';
 import { A4A_MARKETPLACE_LINK } from 'calypso/a8c-for-agencies/components/sidebar-menu/lib/constants';
-import useRemoveSiteMutation from 'calypso/a8c-for-agencies/data/sites/use-remove-site';
 import { urlToSlug } from 'calypso/lib/url/http-utils';
 import { useDispatch, useSelector } from 'calypso/state';
 import { hasAgencyCapability } from 'calypso/state/a8c-for-agencies/agency/selectors';
 import { A4AStore } from 'calypso/state/a8c-for-agencies/types';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
-import { successNotice } from 'calypso/state/notices/actions';
 import isAtomicSite from 'calypso/state/selectors/is-site-automated-transfer';
 import { isJetpackSite } from 'calypso/state/sites/selectors';
 import isA4AClientSite from 'calypso/state/sites/selectors/is-a4a-client-site';
 import { JETPACK_ACTIVITY_ID, JETPACK_BACKUP_ID } from '../features/features';
 import SitesDashboardContext from '../sites-dashboard-context';
 import getActionEventName from './get-action-event-name';
+import createRemoveSiteActionModal from './remove-site-action-modal';
 import type { SiteData } from '../../../../jetpack-cloud/sections/agency-dashboard/sites-overview/types';
 import type { SiteNode, AllowedActionTypes } from '../types';
 
@@ -210,86 +208,10 @@ export default function useSiteActions( {
 
 type SiteActions = {
 	isLargeScreen: boolean;
-	onRefetchSite?: () => Promise< unknown >;
+	onRefetchSite: () => Promise< unknown >;
 	setDataViewsState: ( callback: ( prevState: DataViewsState ) => DataViewsState ) => void;
 	setSelectedSiteFeature: ( siteFeature: string | undefined ) => void;
 };
-
-const createRemoveSiteActionModal =
-	( onRefetchSite: any ) =>
-	( {
-		items,
-		closeModal,
-		onActionPerformed,
-	}: {
-		items: SiteData[];
-		closeModal: () => void;
-		onActionPerformed?: () => void;
-	} ) => {
-		const translate = useTranslate();
-		const dispatch = useDispatch();
-		const { mutate: removeSite } = useRemoveSiteMutation();
-
-		const item = items[ 0 ];
-		const siteName = item.site.value.url;
-		const siteId = item.site.value.a4a_site_id;
-
-		const onRemoveSite = () => {
-			if ( ! siteId ) {
-				return;
-			}
-
-			removeSite(
-				{ siteId },
-				{
-					onSuccess: () => {
-						// Add 1 second delay to refetch sites to give time for site profile to be reindexed properly.
-						setTimeout( () => {
-							onRefetchSite?.()?.then( () => {
-								closeModal?.();
-								onActionPerformed?.();
-								dispatch( successNotice( translate( 'The site has been successfully removed.' ) ) );
-							} );
-						}, 1000 );
-					},
-				}
-			);
-		};
-
-		const onSubmit = ( event: React.FormEvent ) => {
-			event.preventDefault();
-			onRemoveSite();
-		};
-
-		return (
-			<form onSubmit={ onSubmit }>
-				{ translate(
-					'Are you sure you want to remove the site {{b}}%(siteName)s{{/b}} from the dashboard?',
-					{
-						args: { siteName },
-						components: {
-							b: <b />,
-						},
-						comment: '%(siteName)s is the site name',
-					}
-				) }
-				<HStack justify="right">
-					<Button
-						__next40pxDefaultSize
-						variant="tertiary"
-						onClick={ () => {
-							closeModal?.();
-						} }
-					>
-						{ translate( 'Cancel' ) }
-					</Button>
-					<Button __next40pxDefaultSize variant="primary" type="submit" isDestructive>
-						{ translate( 'Remove site' ) }
-					</Button>
-				</HStack>
-			</form>
-		);
-	};
 
 export function useSiteActionsDataViews( {
 	isLargeScreen,
